requirements.txt
Django==4.2.7
Pillow==10.1.0
djangorestframework==3.14.0
pyJWT==2.8.0
django-environ=0.11.2